Use custom scrollbar for group selector for consistent experience
The group selector has some complicated scrolling behavior, and not all browses displayed the scrollbar above the content. Sometimes, placing the content above the scrollbar.

Using css to attempt to alter the scrollbar was not a viable solution due to Firefox not supporting the css required.

Using a custom scrollbar allows us to have control over the location of the scrollbar and get a consistent look accross browsers.