Ensure input event happens in the backdrop before closing
If you were to: mouseDown in a child, followed by dragging your mouse out of the child and a mouseUp in the background, you would receive an onClick and the target would be the backdrop. Then the current code would close the backdrop. This is not good when dragging a scrollbar as it is very easy to be dragging with the mouse down and release within the backdrop, causing accidentally closing.

This also effects touchEvents too. Replace what was said above with touchStart/End instead of mouseDown/Up