Remove extraLarge, set explicit button heights
The old "large" button size was not in the UI specs, and has now been
removed. The former "extraLarge" is now just "large".

There were no existing usages of "extraLarge" anywhere, and all the
existing usages of "large" have been migrated to be "medium".
All of those usages were in the Giving.Web project.

Firefox scales button border widths in unexpected ways:
https://bugzilla.mozilla.org/show_bug.cgi?id=1427391

> Yeah - this is intended behavior (though it can be a bit surprising).
> We snap border-widths to be exact multiples of device pixels, as noted
> in bug 477157 comment 7.

Since we expect the combination of line-height, padding, and border to
exactly equal certain pixel heights, setting that height explicitly is
not really a major change.

Discussion in FLCOM team chat:
https://beta.faithlife.com/messages/12492?threadId=504803