Prevent errors with certain LegacyModal children.
React.Children.only() throws an error if the children prop is not a
single valid React Element. We had previously been checking first if
React.Children.count() was 1 before checking the type, but we had not
been checking whether the children was a valid React Element.

If a consumer tried to pass in a boolean or a string as a child to the
<Modal>, they would be greeted with a page-breaking error in Production.

React.isValidElement() is a much friendlier check that doesn't throw.