chore: Set build status before attempting publish.
Before this change, we were setting "Pending" on the commit being built,
but only setting "Success" on the new "Publish" commit added during the
build. This left the "Pending" status pending forever.

With this new change, we will set the "Success" status when the "Build"
step finishes before attempting to publish. The "Publish" commit will
still get the appropriate "Success" or "Failure" status.