Position calendar and button more precisely
Adjusts the padding on the calendar button so that when the user clicks
above, below, or to the right of the calendar icon but still within the
input field, it activates the calendar button instead of selecting the
input itself.

Also changes the popover reference from the whole button to just the
icon container, so it will be centered correctly if the user specifies a
centered `placement`.