Fix DatePickerInput width handling
Currently, when `width` is given a value less than 100%, it changes the
input field's width but the container remains at 100%, leaving the
calendar button floating on its own way out on the right side of the
container. This fixes that by sending `width` to the container instead
and giving the input field `width="100%"` to match its container.