Update Input and DatePickerInput for v6 (#392).
* Remove styleOverrides prop from Input

* Improve formatting of Input styles

* Replace styleOverrides prop in DatePickerInput

* Remove accidental default variant prop

* Return styleOverrides prop in deprecated status

Reverts commits 4549a5d23a8086fbcf03afb2d4468a62acad8cca and
369dab190cb4f69d884763ec6fe7a49d7f44198a and adds deprecation warnings
and @todo comments.

* Fix default props for DatePickerInput

* Restore legacy component and separate exports

* Add documentation for v6 DatePickerInput

* Remove hyphen from "Styled UI" as per readme title

* Position calendar and button more precisely

Adjusts the padding on the calendar button so that when the user clicks
above, below, or to the right of the calendar icon but still within the
input field, it activates the calendar button instead of selecting the
input itself.

Also changes the popover reference from the whole button to just the
icon container, so it will be centered correctly if the user specifies a
centered `placement`.

* Rename v6 DatePickerInput props to focus on input

This component consists of two basic "things", the input field and the
calendar popover. The input field feels like the main "thing",
positioning-wise, and excess props are already being sent to it, so this
commit makes the following prop name changes to make their purposes more
intuitive:

- inputWidth       -> width
- inputBorderColor -> borderColor
- placement        -> calendarPlacement
- hideShadow       -> hideCalendarShadow
- width            -> calendarWidth
- padding          -> calendarPadding
- border           -> calendarBorder
- zIndex           -> calendarZIndex

* Fix DatePickerInput width handling

Currently, when `width` is given a value less than 100%, it changes the
input field's width but the container remains at 100%, leaving the
calendar button floating on its own way out on the right side of the
container. This fixes that by sending `width` to the container instead
and giving the input field `width="100%"` to match its container.

* Fix CSS comment

* Add v6 import note to DatePickerInput docs

* Fix capitalization of "JSDoc"

* Fix 'validate' prop handling in DatePickerInput

The calendar select popover was rejecting all selected dates if no
'validate' prop was passed.

* Refactor systemPropTypes to styledSystemPropTypes

* Remove calendar-specific props

That's what DatePickerInput.Popover is for.

* Add prop types to child config components

DatePickerInput's, that is.

* Replace "shared-styles" with global theme

* Replace unnecessary ternary operator

* Add DatePickerInput/Input changes to upgrade guide

* Fix accidental plurals

* Move defaultProp to destructuring assignment with default.

* Spread props to Input after other props to allow overriding.

* Remove unnecessary overflow:hidden from documentation.

* Use Box for easy access to the theme.

* Add padding to all sides of calendar icon for better focus highlighting.

Co-authored-by: Robert Bolender <robert.bolender@faithlife.com>
Co-authored-by: Robert Bolender <Robert.Bolender.3@gmail.com>