Fix DatePicker focus-away with min or max dates.
Default browser behavior is to set the document.activeElement to the body
when a previously-focused button becomes programmatically disabled.
Our default focus-away handler on the DatePicker was then closing the
picker immediately after navigating to a month which disabled the active
navigation button.

This PR also makes the DatePicker ever so slightly more navigable with a
keyboard, allowing those buttons to be focused. I'm sure there are
additional ways we can make this component more ARIA-compliant, but I
propose that this first step is at least no worse than before.