Rely on grid and minHeight:0 over height:100%.
Both grid and flexbox give an "implied" minHeight and minWidth of "auto"
to all their immediate children, which has the effect of expanding each
of those elements to the size of all their descendants.

When we had `height:100%` on those descendants, that value was being
calculated based on the height of the top-level grid child, which had
this implied minHeight:auto, which then calculated its height based on
all its descendants, resulting in overflowing the screen.

The workaround that I had previously suggested was to set a maxHeight on
one of the wrapping divs in the Modal.Content component, but that would
require knowing the exact pixel height of the Modal.Header and was a
little bit brittle.

Now that we're specifying minHeight:0 at the top-level, we can remove
the height:100% declarations and rely on grid to expand with their
descendants but still respect the top-level maxHeight based on the
viewport dimensions.