Send appropriate style props to container
When Menu.ActionButton is being used, a lot of Menu.Toggle props are
currently being sent to the caret button that ought to be sent to the
container. For example, one would expect <Menu.Toggle width="100%"> to
only affect the container's width, but right now that would just make
the caret button on the right comically wide.