Send style props to `Menu.Toggle` container when appropriate (#476)
* Add prop types to Menu.Toggle

To help me determine which style props need to be filtered in the next
commit.

* Send appropriate style props to container

When Menu.ActionButton is being used, a lot of Menu.Toggle props are
currently being sent to the caret button that ought to be sent to the
container. For example, one would expect <Menu.Toggle width="100%"> to
only affect the container's width, but right now that would just make
the caret button on the right comically wide.

* Let Menu.ActionButton flex-grow by default

Setting the width of Menu.Toggle doesn't matter if Menu.ActionButton
can't use flex-grow. This can be configured manually without this
commit, but I think this is a more sensible default. For example, one
would expect

<Menu.Toggle width="200px">
  <Menu.ActionButton>Send</Menu.ActionButton>
</Menu.Toggle>

to visually span 200 pixels, but without this commit, you'd have to
write

<Menu.Toggle width="200px" display="inline-flex">
  <Menu.ActionButton flexGrow={1}>Send</Menu.ActionButton>
</Menu.Toggle>