Remove useLegacyChangeHandler
This Hook reverted a breaking change introduced in React Select v3
while Styled UI was still in v5. There's a TODO note on it to remove the
Hook in v6, and we're already warning about this breaking change in the
v6 upgrade guide, so I think it best to remove it now. This isn't
essential to this PR, though.

If this commit is merged, I'll post a special notice in the Styled UI
flow.