Upgrade @babel packages.
Due to changes in preset-env, a couple proposals had to be manually added.

The output from babel was changed because the meaning of the browserlist rule "> 1%" change. "IE 11" was added because that's the only non-modern browser we're likely to want to support.

UglifyJS was removed from the catalog pipeline because it doesn't support the new output.